refactor: refactor codebase for modularity and improved organization
- Refactor codebase to split functionality into multiple files for improved modularity and maintainability
- Move CLI configuration, flag parsing, and validation logic into a dedicated config file
- Move image resizing and processing logic into a separate processor file
- Move batch image processing and file collection logic into its own file
- Move utility functions (file size formatting, logger setup) into a utilities file
- Move version information and version command into a dedicated version file
- Update main initialization to use the new configuration setup
- No functional changes to the core logic, only code organization and separation of concerns
